weatherproject/requirements.txt
pre-commit
black
isort
pytest
pytest-sugar
python-dateutil
django
pyowm
python-dotenv
django-redis
psycopg2
docker
docker-compose
celery